+++ test/node-dependency-injection/Definition.spec.js
@@ -127,4 +127,42 @@ describe('Definition', () => {
       assert.isTrue(definition.lazy)
     })
   })
+
+  describe('deprecated', () => {
+    it('should set as true', () => {
+      // Arrange not needed.
+
+      // Act.
+      definition.deprecated = true
+
+      // Assert.
+      assert.isTrue(definition.deprecated)
+    })
+
+    it('should set as false', () => {
+      // Arrange.
+      definition.deprecated = true
+
+      // Act.
+      definition.deprecated = false
+
+      // Assert.
+      assert.isFalse(definition.deprecated)
+    })
+  })
+
+  describe('setFactory', () => {
+    it('should store the factory with object and method', () => {
+      // Arrange.
+      class Foo {}
+      let method = 'getFactory'
+
+      // Act.
+      definition.setFactory(Foo, method)
+
+      // Assert.
+      assert.strictEqual(definition.factory.Object, Foo)
+      assert.strictEqual(definition.factory.method, method)
+    })
+  })
 })